Add a rendered list of the full stack to PR descriptions
Requires a change to the main update logic: we now need to
always update every PR, since we need to update the messages
of lower PRs when an upper PR is changed.

Also, when creating new PRs, the list of stacked PRs would
be incomplete. Queue all mutations and execute them at the
end to fix this.

Fixes #139

commit-id:28b937ff