Wrap commit body in code area in PR message
Commit messages are not guaranteed to be valid Markdown.
Wrapping them in a code area guarantees that the original
formatting is preserved (say, ASCII art).

This matches what GitHub itself does when rendering commits.

This might be a matter of taste - maybe some people *do*
write valid Markdown in their commits and want to keep the
existing behavior - should we make it a config option?

commit-id:33d44b17