add -r flag for to update command to add reviewers to new PRs
With reference to #145, add a command-line flag to `git spr update` to
add reviewers to any newly created PRs.

    git spr update -r ejoffe

This will, if possible in the context of the repo, request from Eitan
on the newly created PR.

In the case of existing PRs, it will instead print a warning that the
flag has not been applied to a specific pre-existing PR.

    warning: not updating reviewers for PR #1

This is done using the [requestReviews] graphql API call. Since
[requestReviews] operates on the opaque [ID] values, first pull down
the [Repository].`assignableUsers` value and use that to find the [ID]
mapping for each user's login name.

[requestReviews]: https://docs.github.com/en/github-ae@latest/graphql/reference/mutations#requestreviews
[ID]: https://docs.github.com/en/github-ae@latest/graphql/reference/scalars#id
[Repository]: https://docs.github.com/en/github-ae@latest/graphql/reference/objects#repository