avoid multiple "GetAssignableUsers" calls when pushing a new stack
As described in #199, we were calling GetAssignableUsers (a graphql
API query) for each new PR created. This is a property of the
repository, though, and we don't expect it to be changing while we're
pushing a single stack.

Stash the slice of RepoAssignee the first time it's needed and
continue to use it for the rest of the calls. Added a test case to
verify the behaviour.

commit-id:dd9d9f40