git log: Pass --no-abbrev-commit
It's possible to set the `log.abbrevCommit` configuration,
which results in shortened commit hashes:

    log.abbrevCommit
        If true, makes git-log(1), git-show(1), and git-whatchanged(1) assume
        --abbrev-commit. You may override this option with --no-abbrev-commit.

If this configuration is set, spr does not recognize commit hashes
in `git log` because they're not 40 characters long.

This adds the --no-abbrev-commit option to spr's `git log` invocations
so that if a user has that configuration set, it doesn't break spr.