Get rid of space at the end of archive name
For some reason the `make` produced these archives:

```
  • archives
    • creating                                       archive=dist/spr_linux_x86_64 .tar.gz
    • creating                                       archive=dist/spr_linux_arm64 .tar.gz
    • creating                                       archive=dist/spr_windows_x86_64 .zip
    • creating                                       archive=dist/spr_windows_arm64 .zip
    • creating                                       archive=dist/spr_darwin_arm64 .tar.gz
    • creating                                       archive=dist/spr_linux_i386 .tar.gz
    • creating                                       archive=dist/spr_windows_i386 .zip
    • creating                                       archive=dist/spr_darwin_x86_64 .tar.gz
```

Notice the space at the end. I guess it has to do with folding newlines
into whitespaces but for the sake of a quick fix I just reformatted the
`>-` YAML multi-line string as a regular string in one line and now the
problem is gone.

This addresses https://github.com/ejoffe/spr/pull/344#issuecomment-1627388084

Fixup for #340