Enable stricter expectations for github tests.
The application code executes *some* github commands in parallel. This
makes the execution order of those github commands non-deterministic. The existing
unit tests allows for *all* github commands to be executed in any order.
The tests would be clearer and more robust if they enforced a
deterministic execution order. This change updates the tests so they are
executed twice. The first is without goroutines in the application
code so the execution order is deterministic. The second is with
goroutines so the test can ensure that the application code will work
with parallel code.